+++ lily/parse-scm.cc
@@ -39,7 +39,7 @@ internal_ly_parse_scm (Parse_start *ps)
   Source_file *sf = ps->start_location_.get_source_file ();
   SCM port = sf->get_port ();
 
-  int off = ps->start_location_.start () - sf->c_str ();
+  long off = ps->start_location_.start () - sf->c_str ();
 
   scm_seek (port, scm_from_long (off), scm_from_long (SEEK_SET));
   SCM from = scm_ftell (port);
